Moved 4 common functions to utils.py
greedy_policy_from_value_function(), get_policy_map(), greedy_policy_from_value_function() and get_policy_map() moved.
Changed files that used these functions so they work.
