+++ examples/examples.py
@@ -1,8 +1,8 @@
 import numpy as np
 
 from core.envs.gridworld_env import GridWorldEnv
-from core.algorithms.policy_iteration import (single_step_policy_evaluation, reshape_as_gridworld, get_policy_map,
-                                greedy_policy_from_value_function, policy_iteration, value_iteration)
+from core.algorithms import utils
+from core.algorithms.policy_iteration import (policy_iteration, value_iteration)
 
 
 def run_random_gridworld():
@@ -27,35 +27,35 @@ def run_policy_iteration():
     v0 = np.zeros(gw_env.world.size)
     val_fun = v0
     for k in range(500):
-        val_fun = single_step_policy_evaluation(policy0, gw_env, value_function=val_fun)
-    print(reshape_as_gridworld(val_fun, world_shape))
+        val_fun = utils.single_step_policy_evaluation(policy0, gw_env, value_function=val_fun)
+    print(utils.reshape_as_gridworld(val_fun, world_shape))
 
     # test greedy policy
-    policy1 = greedy_policy_from_value_function(policy0, gw_env, val_fun)
-    policy_map1 = get_policy_map(policy1, world_shape)
+    policy1 = utils.greedy_policy_from_value_function(policy0, gw_env, val_fun)
+    policy_map1 = utils.get_policy_map(policy1, world_shape)
     print('Policy: (0=up, 1=right, 2=down, 3=left)\n', policy_map1)
     np.set_printoptions(linewidth=75 * 2, precision=4)
-    print('Policy: (up, right, down, left)\n', get_policy_map(policy1, world_shape))
+    print('Policy: (up, right, down, left)\n', utils.get_policy_map(policy1, world_shape))
     np.set_printoptions(linewidth=75, precision=8)
 
     # test policy iteration
     print('Policy iteration:')
     policy0 = np.ones([gw_env.world.size, len(gw_env.action_state_to_next_state)]) / len(gw_env.action_state_to_next_state)
     optimal_value, optimal_policy = policy_iteration(policy0, gw_env, v0, threshold=0.001, max_steps=1000)
-    print('Value:\n', reshape_as_gridworld(optimal_value, world_shape))
-    print('Policy: (0=up, 1=right, 2=down, 3=left)\n', get_policy_map(optimal_policy, world_shape))
+    print('Value:\n', utils.reshape_as_gridworld(optimal_value, world_shape))
+    print('Policy: (0=up, 1=right, 2=down, 3=left)\n', utils.get_policy_map(optimal_policy, world_shape))
     np.set_printoptions(linewidth=75 * 2, precision=4)
-    print('Policy: (up, right, down, left)\n', get_policy_map(optimal_policy, world_shape))
+    print('Policy: (up, right, down, left)\n', utils.get_policy_map(optimal_policy, world_shape))
     np.set_printoptions(linewidth=75, precision=8)
 
     # test value iteration
     print('Value iteration:')
     policy0 = np.ones([gw_env.world.size, len(gw_env.action_state_to_next_state)]) / len(gw_env.action_state_to_next_state)
     optimal_value, optimal_policy = value_iteration(policy0, gw_env, v0, threshold=0.001, max_steps=100)
-    print('Value:\n', reshape_as_gridworld(optimal_value, world_shape))
-    print('Policy: (0=up, 1=right, 2=down, 3=left)\n', get_policy_map(optimal_policy, world_shape))
+    print('Value:\n', utils.reshape_as_gridworld(optimal_value, world_shape))
+    print('Policy: (0=up, 1=right, 2=down, 3=left)\n', utils.get_policy_map(optimal_policy, world_shape))
     np.set_printoptions(linewidth=75 * 2, precision=4)
-    print('Policy: (up, right, down, left)\n', get_policy_map(optimal_policy, world_shape))
+    print('Policy: (up, right, down, left)\n', utils.get_policy_map(optimal_policy, world_shape))
     np.set_printoptions(linewidth=75, precision=8)
